Update strategy for detecting max offset/bounds.
Instead of calculating the width of the whole slideset, we just
get the first item's width to calculate the 'movedSlides' param
when onTouchEnd is triggered. This is not only simpler, but also
handles the scenario when not all slides have been loaded, which
was making the total 'slidesLength' calculation to fail.